Remote branches also collect author and file information
Get the list of authors and files affected in remote branches.
