+++ app/src/lib/utils/branch.ts
@@ -1,4 +1,4 @@
-import type { CombinedBranch } from "$lib/branches/types";
+import type { CombinedBranch } from '$lib/branches/types';
 
 export function normalizeBranchName(value: string) {
 	return value.replace(/[^A-Za-z0-9_/.#]+/g, '-');
