+++ crates/gitbutler-core/src/virtual_branches/base.rs
@@ -2,15 +2,12 @@ use std::{path::Path, time};
 
 use anyhow::{anyhow, Context, Result};
 use git2::Index;
-use itertools::Itertools;
 use serde::Serialize;
 
 use super::{
-    branch, convert_to_real_branch,
-    integration::{
+    branch, convert_to_real_branch, integration::{
         get_workspace_head, update_gitbutler_integration, GITBUTLER_INTEGRATION_REFERENCE,
-    },
-    target, BranchId, RemoteCommit, VirtualBranchHunk, VirtualBranchesHandle,
+    }, target, BranchId, CommitMetrics, RemoteCommit, VirtualBranchHunk, VirtualBranchesHandle
 };
 use crate::{error::Marker, git::RepositoryExt, rebase::cherry_rebase};
 use crate::{
@@ -20,14 +17,6 @@ use crate::{
     users,
     virtual_branches::branch::BranchOwnershipClaims,
 };
-
-#[derive(Debug, Serialize, PartialEq, Clone)]
-#[serde(rename_all = "camelCase")]
-pub struct CommitMetrics {
-    pub name: String,
-    pub value: u32,
-    pub commit_ids: Vec<String>,
-}
 
 #[derive(Debug, Serialize, PartialEq, Clone)]
 #[serde(rename_all = "camelCase")]
@@ -334,52 +323,6 @@ fn _print_tree(repo: &git2::Repository, tree: &git2::Tree) -> Result<()> {
     Ok(())
 }
 
-fn get_recent_commits_metric(
-    recent_commits: &Vec<RemoteCommit>,
-) -> (Vec<CommitMetrics>, Vec<CommitMetrics>) {
-    let mut recent_authors =
-        std::collections::HashMap::<String, CommitMetrics>::with_capacity(recent_commits.len());
-    let mut recent_files = std::collections::HashMap::<String, CommitMetrics>::new();
-    for commit in recent_commits {
-        let commit_id = &commit.id;
-        let author_name = &commit.author.name;
-        let author_metric =
-            recent_authors
-                .entry(author_name.to_string())
-                .or_insert(CommitMetrics {
-                    name: author_name.to_string(),
-                    value: 0,
-                    commit_ids: Vec::new(),
-                });
-        author_metric.value += 1;
-        author_metric.commit_ids.push(commit_id.to_string());
-        for file_path in &commit.file_paths {
-            let file_metric = recent_files
-                .entry(file_path.to_string())
-                .or_insert(CommitMetrics {
-                    name: file_path.to_string(),
-                    value: 0,
-                    commit_ids: Vec::new(),
-                });
-            file_metric.value += 1;
-            file_metric.commit_ids.push(commit_id.to_string());
-        }
-    }
-
-    let recent_authors = recent_authors
-        .into_values()
-        .sorted_by_key(|a| a.value)
-        .rev()
-        .collect();
-
-    let recent_files = recent_files
-        .into_values()
-        .sorted_by_key(|f| f.value)
-        .rev()
-        .collect();
-
-    (recent_authors, recent_files)
-}
 
 // try to update the target branch
 // this means that we need to:
@@ -663,7 +606,7 @@ pub fn target_to_base_branch(
         .collect::<Vec<_>>();
 
     // get recent authors and file paths, and their appearance count
-    let (recent_authors, recent_files) = get_recent_commits_metric(&recent_commits);
+    let (recent_authors, recent_files) = super::get_recent_commits_metric(&recent_commits);
 
     // there has got to be a better way to do this.
     let push_remote_url = match target.push_remote_name {
