+++ crates/gitbutler-core/src/virtual_branches/mod.rs
@@ -23,3 +23,6 @@ pub use remote::*;
 mod state;
 pub use state::VirtualBranches as VirtualBranchesState;
 pub use state::VirtualBranchesHandle;
+
+mod metrics;
+pub use metrics::*;
